backup_v3_selenium/requirements.txt
mysql-connector-python
requests
schedule
selenium
webdriver-manager
selenium
webdriver-manager
python-dotenv
pyperclip
